# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AfroHit Studio API
  description: |
    Public REST contract. All endpoints under `/api/v1` require a Clerk bearer JWT.
    The Studio Chat endpoint accepts free-form prompts and orchestrates the labs by
    calling internal tools (hooks, lyrics, beats, vocals, etc.).
  version: 0.1.0
servers:
  - url: http://localhost:4000
    description: local
  - url: https://api.afrohit.studio
    description: production
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatMessage:
      type: object
      required: [content]
      properties:
        threadId: { type: string, nullable: true }
        projectId: { type: string, nullable: true }
        content: { type: string, minLength: 1, maxLength: 8000 }
...
paths:
...
  /api/v1/projects/{projectId}/briefs:
    post:
      summary: Create a brief
      parameters: [{ in: path, name: projectId, required: true, schema: { type: string } }]
      responses: { '201': { description: created } }
  /api/v1/projects/{projectId}/briefs/polish:
    post:
      summary: AI-polish a free-form idea into a brief
      parameters: [{ in: path, name: projectId, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [rawIdea], properties: { rawIdea: { type: string } } } } }
      responses: { '201': { description: created } }
...
  /api/v1/chat/messages:
    post:
      summary: Send a chat message (orchestrates labs via tool calls)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChatMessage' } } }
      responses: { '200': { description: ok } }
  /api/v1/chat/messages/stream:
    post:
      summary: Streaming chat (Server-Sent Events — thread/tool_start/tool_result/assistant/done)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChatMessage' } } }
      responses: { '200': { description: text/event-stream } }
...
  /api/v1/admin/stats:
    get: { summary: Operator stats (gated by ADMIN_EMAILS), responses: { '200': { description: ok }, '403': { description: not an operator } } }
...
  /api/v1/admin/workspaces/{id}/credits:
    post: { summary: Grant/claw back credits (operator), parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /api/v1/admin/workspaces/{id}/suspend:
    post: { summary: Suspend workspace (operator), parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /api/v1/admin/workspaces/{id}/unsuspend:
    post: { summary: Unsuspend workspace (operator), parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
...
  /api/v1/admin/jobs/{id}/retry:
    post: { summary: Re-enqueue a failed job (operator), parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
...